refactored lift, slide intake and servo class to individual files and refactored auto opmode to use these classes.
